requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.23
kombu==4.6.4
more-itertools==7.2.0
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.5
vine==1.3.0
zipp==0.6.0
